Fix android coordinates (#4116)
* chore: example for testing screen coordinates

* fix(android): screen coordinate conversion on android

* fix(ios): avoid loss of precision Double -> Float -> CGFloat

* fix: improve docs

* chore: remove pixel conversion comments

* fix: pixel conversion comments

+++ ios/RNMBX/RNMBXMapViewModule.mm
@@ -60,7 +60,7 @@ RCT_EXPORT_METHOD(getCoordinateFromView:(nonnull NSNumber*)viewRef atPoint:(NSAr
         NSNumber* a = [atPoint objectAtIndex:0];
         NSNumber* b = [atPoint objectAtIndex:1];
 
-        [view getCoordinateFromViewWithAtPoint:CGPointMake(a.floatValue, b.floatValue) resolver:resolve rejecter:reject];
+        [view getCoordinateFromViewWithAtPoint:CGPointMake(a.doubleValue, b.doubleValue) resolver:resolve rejecter:reject];
     } reject:reject methodName:@"getCoordinateFromView"];
 }
